requirements.txt
fastapi
uvicorn
python-multipart
pydantic
spacy
pdfplumber
python-docx
sentence-transformers
matplotlib
seaborn
notebook
scikit-learn
filetype
pytesseract
pillow
optimum[openvino,nncf]
umap-learn
hdbscan